refactor bibliotheque
