+++ chat/commun/noeudChat.ts
@@ -1,4 +1,12 @@
-import { FormatNoeudMutable, NoeudMutable, NoeudEnveloppeMutable, creerCentreSansVoisins, EtiquetteNoeud, FormatNoeudImmutable, NoeudImmutable, NoeudEnveloppeImmutable } from "../../bibliotheque/communication";
+import {Sommet} from "../../bibliotheque/communication/communication";
+
+import {
+    NoeudEnveloppeMutable, NoeudEnveloppeImmutable, NoeudMutable, NoeudImmutable, FormatNoeudImmutable, FormatNoeudMutable, EtiquetteNoeud,
+} from "../../bibliotheque/communication/noeud";
+
+import {creerCentreSansVoisins,} from "../../bibliotheque/communication/creerReseau";
+
+
 import { creerTableImmutable } from "../../bibliotheque/types";
 import { jamais } from "../../bibliotheque/outils";
 import {FormatSommetChat, creerSommetChat} from './sommetChat';
