+++ chat/commun/reseauChat.ts
@@ -1,6 +1,16 @@
 import {
-    Sommet, ReseauMutable, AssemblageReseau,
-    creerAssemblageReseauEnAnneau} from "../../bibliotheque/communication";
+    Sommet
+} from "../../bibliotheque/communication/communication";
+
+
+import {ReseauMutable} from "../../bibliotheque/communication/creerReseau";
+
+import {
+    creerAssemblageReseauEnAnneau, AssemblageReseau
+} from "../../bibliotheque/communication/assemblageReseau";
+
+
+
 import {Identification, creerIdentificationParCompteur} from "../../bibliotheque/types";
 import { jamais } from "../../bibliotheque/outils";
 import {FormatSommetChat} from "./sommetChat";
